Dim iosdbg prompt and other improvements

+++ source/memutils.h
@@ -17,6 +17,6 @@ kern_return_t memutils_dump_memory_new(unsigned long long location, vm_size_t am
 kern_return_t memutils_read_memory_at_location(void *, void *, vm_size_t);
 kern_return_t memutils_write_memory_to_location(vm_address_t, vm_offset_t);
 long long memutils_buffer_to_number(void *, int);
-kern_return_t memutils_valid_location(unsigned long long);
+kern_return_t memutils_valid_location(unsigned long);
 
 #endif
